refactor: general cleanup to facilitate feature development
